feat : 조회수 캐싱 동시성 처리, 스케줄링 변경 RedisViewCountService : 동시성 문제 방지를 위한 동기화 코드 추가. ViewCountSyncScheduler : DB 동기화 스케줄링 10초 -> 60초로 변경
refactor : 로직 분리
Redis 적용 전, 적용 후 컨트롤러, 서비스, 리포지토리 분리
RedisKeyHelper uitl 패키지로 이동

+++ src/main/java/com/github/mingyu/bigboard/dto/BoardScore.java
@@ -11,11 +11,11 @@ import lombok.Setter;
 @NoArgsConstructor
 public class BoardScore {
     private Long boardId;
-    private Double score= 0.0;
-    private Integer totalScore = 0;
-    private Integer ratingCount = 0;
+    private double score= 0.0;
+    private double totalScore = 0.0;
+    private int ratingCount = 0;
 
     public double getAverageScore() {
-        return ratingCount == 0 ? 0 : (double) totalScore / ratingCount;
+        return ratingCount == 0 ? 0 : totalScore / ratingCount;
     }
 }
